requirements.txt
openai
langchain
langgraph
faiss-cpu
pdfplumber
streamlit
tiktoken
pypdf
python-dotenv
requests
langchain_community